taskpanes: refactor all taskpanes to be better components
