+++ mitosheet/src/components/taskpanes/Search/SearchTaskpane.tsx
@@ -2,10 +2,12 @@
 // Distributed under the terms of the Modified BSD License.
 
 import React, { useEffect } from 'react';
-import DefaultTaskpane from '../DefaultTaskpane';
+import DefaultTaskpane from '../DefaultTaskpane/DefaultTaskpane';
 import MitoAPI from '../../../api';
 import { GridState, SheetData, UIState } from '../../../types';
 import Search from './Search';
+import DefaultTaskpaneHeader from '../DefaultTaskpane/DefaultTaskpaneHeader';
+import DefaultTaskpaneBody from '../DefaultTaskpane/DefaultTaskpaneBody';
 
 interface SearchTaskpaneProps {
     sheetData: SheetData | undefined;
@@ -35,10 +37,12 @@ const SearchTaskpane = (props: SearchTaskpaneProps): JSX.Element => {
     }, [])
 
     return (
-        <DefaultTaskpane
-            header='Search Values in Sheet'
-            setUIState={props.setUIState}   
-            taskpaneBody={
+        <DefaultTaskpane>
+            <DefaultTaskpaneHeader
+                header='Search Values in Sheet'
+                setUIState={props.setUIState}
+            />
+            <DefaultTaskpaneBody>
                 <Search
                     gridState={props.gridState}
                     sheetData={props.sheetData}
@@ -47,8 +51,8 @@ const SearchTaskpane = (props: SearchTaskpaneProps): JSX.Element => {
                     mitoContainerRef={props.mitoContainerRef}
                     uiState={props.uiState}
                 />
-            }     
-        />            
+            </DefaultTaskpaneBody>
+        </DefaultTaskpane>     
     )
 };
